requirements_api.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-telegram-bot>=20.0
httpx>=0.25.0
requests>=2.31.0
matplotlib>=3.7.0
pandas>=2.0.0
schedule>=1.2.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1